+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpVerifier.kt
@@ -385,6 +385,7 @@ class OpenId4VpVerifier(
         authenticationRequestParameters,
     )
 
+    @Suppress("DEPRECATION")
     private fun RequestOptions.clientMetadata(): RelyingPartyMetadata? = when (clientIdScheme) {
         is ClientIdScheme.RedirectUri,
         is ClientIdScheme.VerifierAttestation,