+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/RequestOptions.kt
@@ -72,9 +72,7 @@ data class RequestOptions(
     /** Opaque value which will be returned by the OpenId Provider and also in [AuthnResponseResult]. */
     val state: String = uuid4().toString(),
 
-    /**
-     * Set this value to include metadata with encryption parameters set.
-     */
+    @Deprecated("Encryption depends on [responseMode]")
     val encryption: Boolean = false,
 
     /**